split SVG generating class into two
