+++ .vscode/settings.json
@@ -1,4 +1,5 @@
 {
     "java.configuration.updateBuildConfiguration": "interactive",
-    "files.eol": "\n"
+    "files.eol": "\n",
+    "java.compile.nullAnalysis.mode": "automatic"
 }
