+++ app/src/main/java/qrcode/svg/example/App.java
@@ -9,8 +9,10 @@ public class App {
         try {
             // generate a PNG format QR Code file
             new PrintQR().write();
-            // generate an SVG format QR Code file
-            new PrintQRSVG().write();
+            // generate an SVG format QR Code file (square)
+            new PrintQRSVGSquare().write();
+            // generate an SVG format QR Code file (circle)
+            new PrintQRSVGCircle().write();
         } catch(Exception e) {
             e.printStackTrace();
         }
