+++ app/src/main/java/qrcode/svg/example/PrintQR.java
@@ -54,7 +54,5 @@ public class PrintQR {
         BitMatrix bm = qw.encode(content, BarcodeFormat.QR_CODE, canvasSize, canvasSize, hints);
         BufferedImage img = MatrixToImageWriter.toBufferedImage(bm);
         ImageIO.write(img, "png", new File(fileOutputPath + outputFileName));
-
-        
     }
 }
